# docs/swagger.yaml
openapi: 3.0.0
info:
    title: ZideQuest
    version: 1.0.0
servers:
    - url: https://feline-precious-wildcat.ngrok-free.app/api
    - url: http://localhost:3000/api
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
...
paths:
...
    /v1/users/quests:
        get:
            tags:
                - user
            summary: get users quest
            security:
                - bearerAuth: []
            responses:
                "200":
                    description: Successful response
                    content:
                        application/json: {}
...
    /v1/quests/{id}:
        put:
            tags:
                - quest
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
            summary: update quest
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            example: ""
            security:
                - bearerAuth: []
            responses:
                "200":
                    description: Successful response
                    content:
                        application/json: {}
        delete:
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
            tags:
                - quest
            summary: delete quest
            security:
                - bearerAuth: []
            responses:
                "200":
                    description: Successful response
                    content:
                        application/json: {}
...